pocs/typesafe-llm-pydantic-ai/requirements.txt
pydantic-ai-slim[google]>=0.0.18
python-dotenv>=1.0.0
